Rename files for readability.
